refactor(gfql): extern carried-output resolution out of gfql_unified, mirror reentry tests
Externs the optional-reentry carried-output symbols out of the oversized
gfql_unified.py into graphistry/compute/gfql/cypher/reentry/carried_outputs.py,
replaces the getattr AST probes with isinstance dispatch on ASTCall, and shares
one identifier predicate across flatten/carry/carried_outputs via
graphistry.compute.gfql.identifiers.

Tests pinning cypher/reentry/* now mirror the source tree under
graphistry/tests/compute/gfql/cypher/reentry/.
