+++ graphistry/compute/gfql_unified.py
@@ -32,11 +32,6 @@ from graphistry.compute.gfql.same_path_types import (
     parse_where_json,
 )
 from graphistry.compute.exceptions import ErrorCode, GFQLValidationError
-from graphistry.compute.gfql.agg_types import (
-    CYPHER_EMPTY_LIST_EMPTY_GROUP_AGGREGATIONS,
-    CYPHER_ZERO_EMPTY_GROUP_AGGREGATIONS,
-    CypherEmptyGroupValue,
-)
 from graphistry.compute.gfql.cypher.ast import CypherParams
 from graphistry.compute.gfql.cypher.parser import parse_cypher
 from graphistry.compute.gfql.exec_context import attach_row_exec_context, clear_row_exec_context
@@ -50,9 +45,11 @@ from graphistry.compute.gfql.cypher.lowering import (
     compile_cypher_query,
 )
 from graphistry.compute.filter_by_dict import _node_dtypes_for_pushdown
+from graphistry.compute.gfql.cypher.reentry.carried_outputs import (
+    carried_output_sources as _carried_output_sources,
+    optional_reentry_aggregate_fill_values as _optional_reentry_aggregate_fill_values,
+)
 from graphistry.compute.gfql.cypher.reentry.execution import (
-    CARRIED_OUTPUTS_NOT_REPRODUCIBLE,
-    CarriedOutputSources,
     REENTRY_DUPLICATE_CARRIED_ROWS_REASON as _REENTRY_DUPLICATE_CARRIED_ROWS_REASON,
     REENTRY_WHOLE_ROW_SUGGESTION as _REENTRY_WHOLE_ROW_SUGGESTION,
     apply_optional_reentry_null_fill as _apply_optional_reentry_null_fill,
@@ -63,7 +60,6 @@ from graphistry.compute.gfql.cypher.reentry.execution import (
     reentry_validation_error as _reentry_validation_error,
     union_scalar_reentry_results as _union_scalar_reentry_results,
 )
-from graphistry.compute.gfql.cypher.reentry.naming import is_reentry_hidden_column_reference
 from graphistry.compute.gfql.cypher.call_procedures import execute_cypher_call
 from graphistry.compute.gfql.cypher.result_postprocess import (
     apply_result_projection,
@@ -1509,121 +1505,6 @@ def _execute_compiled_query_with_reentry(
         )
 
     return result
-
-
-_IDENTIFIER_TOKEN = re.compile(r"[A-Za-z_][A-Za-z0-9_]*")
-_REENTRY_MARKER_COLUMN = re.compile(r"__cypher_reentry_(\w+)__")
-
-
-def _output_reads_carried_alias(
-    src: str, *, alias_names: Set[str], scalar_columns: Set[str]
-) -> bool:
-    tokens = set(_IDENTIFIER_TOKEN.findall(src))
-    return bool(tokens & alias_names or tokens & scalar_columns)
-
-
-def _carried_output_source_column(
-    src: str, *, alias_names: Set[str], scalar_columns: Set[str]
-) -> Optional[str]:
-    """Prefix-frame column an output copies verbatim, or None when it cannot be reproduced."""
-    if src in scalar_columns:
-        return src
-    parts = src.split(".")
-    if len(parts) != 2 or parts[0] not in alias_names:
-        return None
-    marker = _REENTRY_MARKER_COLUMN.fullmatch(parts[1])
-    if marker is not None and marker.group(1) in scalar_columns:
-        return marker.group(1)
-    if _IDENTIFIER_TOKEN.fullmatch(parts[1]):
-        return src
-    return None
-
-
-def _carried_output_sources(compiled_query: CompiledCypherQuery) -> CarriedOutputSources:
-    """Which result outputs the unmatched-prefix-row null-fill can reproduce (#1896)."""
-    plan = compiled_query.reentry_plan
-    if plan is None:
-        return CarriedOutputSources(columns={}, every_output_reproducible=True)
-    alias_names = {plan.reentry_alias_name} | {a.output_name for a in plan.aliases}
-    scalar_columns = set(plan.scalar_columns)
-    items: Optional[List[object]] = None
-    grouped = False
-    ops = list(compiled_query.chain.chain) if compiled_query.chain is not None else []
-    for op in ops:
-        function = getattr(op, "function", None)
-        op_params = getattr(op, "params", None) or {}
-        if function in ("select", "return_", "with_") and op_params.get("items"):
-            items = list(op_params["items"])
-            grouped = False
-        elif function == "group_by":
-            grouped = True
-    if items is None or grouped:
-        return CarriedOutputSources(columns={}, every_output_reproducible=True)
-    columns: Dict[str, str] = {}
-    for item in items:
-        if not (isinstance(item, (list, tuple)) and len(item) == 2 and isinstance(item[0], str)):
-            continue
-        name, src = item[0], item[1]
-        if not isinstance(src, str):
-            continue
-        if not _output_reads_carried_alias(
-            src, alias_names=alias_names, scalar_columns=scalar_columns
-        ):
-            continue
-        source_column = _carried_output_source_column(
-            src, alias_names=alias_names, scalar_columns=scalar_columns
-        )
-        if source_column is None:
-            return CARRIED_OUTPUTS_NOT_REPRODUCIBLE
-        columns[name] = source_column
-    return CarriedOutputSources(columns=columns, every_output_reproducible=True)
-
-
-def _optional_reentry_aggregate_fill_values(compiled_query: CompiledCypherQuery) -> Dict[str, CypherEmptyGroupValue]:
-    """Cypher empty-group value per aggregate output on an unmatched prefix row's null-extended row."""
-    plan = compiled_query.reentry_plan
-    carried_scalar_columns = set(plan.scalar_columns) if plan is not None else set()
-    reentry_alias = plan.reentry_alias_name if plan is not None else None
-
-    def source_is_carried_rather_than_suffix_bound(source: str) -> bool:
-        base = source.split(".", 1)[0]
-        return (
-            is_reentry_hidden_column_reference(source)
-            or base in carried_scalar_columns
-            or base == reentry_alias
-        )
-
-    ops = list(compiled_query.chain.chain) if compiled_query.chain is not None else []
-    with_map: Dict[str, object] = {}
-    fills: Dict[str, CypherEmptyGroupValue] = {}
-    for op in ops:
-        if not isinstance(op, ASTCall):
-            continue
-        if op.function == "with_":
-            for item in op.params.get("items") or []:
-                if isinstance(item, (list, tuple)) and len(item) == 2 and isinstance(item[0], str):
-                    with_map[item[0]] = item[1]
-        elif op.function == "group_by":
-            fills = {}
-            for agg in op.params.get("aggregations") or []:
-                if not isinstance(agg, (list, tuple)) or len(agg) not in (2, 3):
-                    continue
-                alias = str(agg[0])
-                func = str(agg[1]).lower()
-                expr = agg[2] if len(agg) == 3 else None
-                if func == "count" and (expr is None or expr == "*"):
-                    fills[alias] = 1
-                    continue
-                source: object = with_map.get(expr, expr) if isinstance(expr, str) else expr
-                if not isinstance(source, str):
-                    continue
-                if source_is_carried_rather_than_suffix_bound(source):
-                    continue
-                if func in CYPHER_ZERO_EMPTY_GROUP_AGGREGATIONS:
-                    fills[alias] = 0
-                elif func in CYPHER_EMPTY_LIST_EMPTY_GROUP_AGGREGATIONS:
-                    fills[alias] = []
-    return fills
 
 
 def _materialize_split_alias_columns(
